Move Convert to separate file, fix readme & Info
