+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/key.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Germanikus</title>
+    <title>Germanikus Cipher</title>
   </head>
   <body>
     <div id="app"></div>
